move inference dataset to the inference submodule
